refactor: strata naming in parking-lot-shared
Dissolve graph.ts into the strata convention the framework README
documents: tags at the bottom, atom.store/resource.tx as nouns,
flow.* files as verbs, imports flowing downward only. Pure helper
functions stay unprefixed in rules.ts — prefixes mark graph nodes,
plain modules stay plain. Public index surface unchanged.
