+++ examples/parking-lot-shared/src/atom.store.ts
@@ -1,0 +1,6 @@
+import { atom } from "@pumped-fn/lite"
+import { createMemoryStore } from "./store"
+
+export const store = atom({
+  factory: () => createMemoryStore(),
+})
